Adding more flexibility to input for interactable objects
Added new class that separates interactable objects with an instant interaction and interactable objects with the need to hold down the button for a certain amount of time. This time can be tweaked an adjsuted. ALos adjusted the player interaction key triggers to acomodate for this along with adding new functions to the player to deal with dertermining which type of interaction trigger to use when the controller receives input based on if the current interactable object reference inherits from the interactableObjectHold class or the InteractableObject class.